fix(openclaw): emit channel user message to both IPC streams for AionUI rendering
Move user message persistence and IPC emission from ChannelMessageService
into OpenClawAgentManager.sendChannelMessage. Emit user_content to both
openclawConversation and conversation response streams, matching the
pattern used by handleStreamEvent for AI responses. Pass msg_id from
ChannelMessageService to sendChannelMessage for proper message tracking.
